[arcane,accelerator] Déplace la définition de 'ReduceMemoryImpl' dans son propre fichier d'en-tête.
